feat: Replaced interface with abstract class with common behavior
